created new script for handling custom predicates combinations. renames suggestions and nits also solved.
